refactor: restructure dataclass src and tests modules
- Consolidated and reorganized test modules for dataclasses, splitting them into logical subdirectories for better maintainability and readability.
- Introduced new tests for `SensorMsgsImu` to ensure coverage for added functionality.
- Updated imports and feature mappings to align with the revised structure and improved modularity.

Issue TCT-52
