+++ src/trajectory_container_tools/dataclasses/ros_msgs/nested_dataclass.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 
 import numpy as np
 
-from .core_dataclass import NestedRosStampedDataclass, RosStampedDataclass, NestedBaseTrajectoryDataclass
+from .core_dataclass import NestedRosStampedDataclass, NestedBaseTrajectoryDataclass
 from trajectory_container_tools.dataclasses.ros_msgs.primitive_dataclass import Point, \
     Quaternion, Transform, Vector3
 
@@ -123,42 +123,6 @@ class AckermannMsgsAckermannDrive(NestedBaseTrajectoryDataclass):
 
 
 @dataclass()
-class SensorMsgsImu(RosStampedDataclass):
-    """Represents IMU (Inertial Measurement Unit) sensor data with orientation,
-    angular velocity, and
-    linear acceleration, including their covariance values.
-
-    Compatible ros2 message interface: sensor_msgs/msg/Imu
-
-    This dataclass is used to store data typically obtained from an IMU sensor. It contains
-    information about the orientation, angular velocity, and linear acceleration of a body, along
-    with their respective covariance matrices to represent variability or uncertainty in
-    measurements.
-
-    :ivar orientation: The orientation of the sensor expressed as a quaternion.
-    :type orientation: Quaternion
-    :ivar orientationCovariance: The covariance matrix of the orientation measurement.
-    :type orientationCovariance: np.ndarray
-    :ivar angularVelocity: The angular velocity of the sensor.
-    :type angularVelocity: Vector3
-    :ivar angularVelocityCovariance: The covariance matrix of the angular velocity measurement.
-    :type angularVelocityCovariance: np.ndarray
-    :ivar linearAcceleration: The linear acceleration of the sensor.
-    :type linearAcceleration: Vector3
-    :ivar linearAccelerationCovariance: The covariance matrix of the linear acceleration
-    measurement.
-    :type linearAccelerationCovariance: np.ndarray
-    """
-
-    orientation: Quaternion
-    orientationCovariance: np.ndarray
-    angularVelocity: Vector3
-    angularVelocityCovariance: np.ndarray
-    linearAcceleration: Vector3
-    linearAccelerationCovariance: np.ndarray
-
-
-@dataclass()
 class VescMsgsVescImu(NestedBaseTrajectoryDataclass):
     """
     Represents IMU data related to VESC (Vedder Electronic Speed Controller).
@@ -187,7 +151,7 @@ class VescMsgsVescImu(NestedBaseTrajectoryDataclass):
     compass: Vector3
     orientation: Quaternion
 
-
+# ==== Nested and stamped =========================================================================
 @dataclass()
 class TransformStamped(NestedRosStampedDataclass):
     # Compatible ros2 message interface: geometry_msgs/msg/TransformStamped
